slateo-api/requirements.txt
allauth-django==0.42.6
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.3
django-cors-headers==3.6.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
dnspython==2.0.0
filelock==3.0.12
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.3
oauthlib==3.1.0
openapi-codec==1.3.2
phonenumbers==8.12.15
Pillow==8.1.0
py3-validate-email==0.2.12
pycparser==2.20
Pygments==2.7.2
PySocks==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
